Tidy up vertical landing pages
